requirement.txt
asgiref==3.5.1
Django==4.0.4
flake8==4.0.1
flake8-html==0.4.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
Pillow==9.1.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
sqlparse==0.4.2
